requirements.txt
# Project dependencies
fastmcp>=0.1.0
python-dotenv>=1.0.0
mysql-connector-python>=8.0.0 